Incrementally populate transient--suffixes
This allows suffixes to access the up-to-date values of earlier
suffixes.  Doing so was already possible by instead using the value of
the prefix's `value' slot, which at this time often isn't outdated yet.

And yet, the prefix's `value' is not certain not to be outdated, e.g.,
when a suffix has a non-nil `init-value'.  So it is often better to get
the up-to-date value using `transient-get-value', which thanks to this
commit is now possible even during the initial setup of the menu.

But suffixes do have to be processed in some order, and we obviously
cannot use the result of an operation that hasn't taken place yet.
So to use `transient-get-value' at this time, suffixes have to be
ordered so that the value of suffixes whose values are used by other
suffixes are processes earlier.  They are processed earlier if they
appear earlier (further up and more to the left) in the menu.

In some rare cases that may not be possible or desirable.  In such
cases it may still make sense to use the value of the prefix's `value'
slot.

Re #316.