Prevent some misguided access of a prefix's value slot
Initializing a prefix sets its `value' slot.  Initializing the `value'
slots of its suffixes involves accessing the value stored in that slot.
After that the values of the suffixes' `value' slot collectively specify
the prefix's value.  The value of the prefix's `value' slot may or may
not be identical at this point.  (For example a suffix's `init-value'
slot may set pick a value different from the value in the prefix's
`value' slot.)

To prevent access to the potentially outdated value of the prefix's
`value' slot, make it unbound after initializing the suffixes.