transient--init-suffix: No longer drop suffix if it is inactive
Each suffix is used to track its value.  If we drop an inactive
suffix and it later becomes active, then its value is lost, so
stop doing that.  The consequence is that we now have to account
for `inactive' suffixes in several places, just like we already
had to do for `inapt' suffixes.

A suffix may be inactive for two reason.  It may have an `:if*'
predicate, which currently returns nil.  Or it may be on a level
that is currently not being used.

Closes #316.
Closes #364.
Closes #407.