transient--get-savable-value: Fall back to transient-current-suffixes
When we created `transient--get-savable-value' in [1: b4edb633], we
(intentionally) stopped falling back to `transient-current-suffixes'.
The value is saved to history before `transient--suffixes' is unset
and our commands that explicitly set the value do not exit the menu.

However, a third-party may add a command that sets the value and
also exits the menu.  I realized that when I made the above change,
but figured I would hear about it, if anybody actually did that.
That just happened with #410, so use `transient-current-suffixes'
again.

Closes #410.

1: 2025-08-16 b4edb633488fc251df0883b69b1eb350cb2e94ec
   transient-get-value: Replace with three functions