Resume transient when isearch is exited using non-isearch binding
An event that has no binding in `isearch-mode-map' causes Isearch to
be exited (see `isearch-pre-command-hook').  When that happens we have
to resume the transient menu just like when the user explicitly uses a
a command whose purpose it is to exit Isearch.

We deal with regular Isearch exit commands by rebinding them to wrapper
commands in our `isearch-mode-map' wrapper.  We can deal with irregular
isearch exits by using one of our wrapper commands as the default
binding in that keymap.