transient--pp-to-file: Avoid saving empty value
Do not create the file and containing directory if the value to be
saved is nil.  If the file already exists, then always update, even
if the value is nil, to ensure an older value doesn't stick around.

If someone customized where history is stored, but has to use "emacs
-Q", then they used to end up with "~/.emacs.d/transient/history.el"
with content "nil".